moves some extensions to only file they're used
