--- data-collection/data-collection/Extensions/UIKit/UIViewController/UIViewController+Root.swift
@@ -1,24 +1,0 @@
-//// Copyright 2018 Esri
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import UIKit
-
-extension UIViewController {
-    
-    /// Determine if the view controller is the first view controller (root) in the navigation controller's view controller stack.
-    ///
-    var isRootViewController: Bool {
-        return self == navigationController?.viewControllers.first
-    }
-}
