+++ data-collection/data-collection/View Controllers/Rich Popup View Controllers/Rich Popup View Controller/RichPopupViewController.swift
@@ -421,3 +421,26 @@ class RichPopupViewController: SegmentedViewController {
     
     var isProcessingNewAttachmentImage: Bool = false
 }
+
+fileprivate extension UIViewController {
+    
+    func popOrDismiss(animated: Bool, completion: (() -> Void)? = nil) {
+        
+        if let navigationController = navigationController {
+            
+            if isRootViewController {
+                dismiss(animated: animated, completion: completion)
+            }
+            else {
+                navigationController.popViewController(animated: animated, completion: completion)
+            }
+        }
+        else {
+            dismiss(animated: animated, completion: completion)
+        }
+    }
+    
+    var isRootViewController: Bool {
+        return self == navigationController?.viewControllers.first
+    }
+}
